refactor: rewrite config-manager in TS
- autocomplete friendly
- works with browser
