+++ packages/config-manager/package.json
@@ -5,8 +5,7 @@
   "author": "Xuejie Xiao <xxuejie@gmail.com>",
   "homepage": "https://github.com/nervosnetwork/lumos#readme",
   "license": "MIT",
-  "main": "lib/index.js",
-  "types": "index.d.ts",
+  "main": "lib",
   "engines": {
     "node": ">=12.0.0"
   },
@@ -32,14 +31,19 @@
     "url": "git+https://github.com/nervosnetwork/lumos.git"
   },
   "scripts": {
-    "fmt": "prettier --write \"{lib,tests}/**/*.js\" index.d.ts package.json",
-    "lint": "eslint -c ../../.eslintrc.js \"{lib,tests}/**/*.js\"",
-    "test": "ava"
+    "fmt": "prettier --write \"{src,tests}/**/*.ts\" package.json",
+    "lint": "eslint -c ../../.eslintrc.js \"{src,tests}/**/*.ts\"",
+    "test": "ava",
+    "build": "tsc",
+    "clean": "rm -rf lib",
+    "prepublishOnly": "yarn run clean && yarn run build",
+    "release": "npm publish"
   },
   "bugs": {
     "url": "https://github.com/nervosnetwork/lumos/issues"
   },
   "dependencies": {
+    "@types/deep-freeze-strict": "^1.1.0",
     "deep-freeze-strict": "^1.1.1"
   }
 }
