+++ packages/config-manager/src/index.ts
@@ -1,0 +1,4 @@
+export * from "./types";
+export { initializeConfig, getConfig, validateConfig } from "./manager";
+export * as helpers from "./helpers";
+export { predefined, createConfig } from "./predefined";
